refactor: split native video tests into with style and without
