+++ packages/video/__tests__/android/video-with-style.android.test.js
@@ -1,0 +1,3 @@
+import shared from "../video-shared-with-style.native";
+
+shared();
