+++ packages/video/__tests__/ios/video-with-style.ios.test.js
@@ -1,0 +1,3 @@
+import shared from "../video-shared-with-style.native";
+
+shared();
